+++ CHANGES.md
@@ -15,6 +15,7 @@ Change Log
 * Catalog items can now show a disclaimer or message before loading through specifying `InitialMessageTraits`
 * Added Leaflet hack to remove white-gaps between tiles (https://github.com/Leaflet/Leaflet/issues/3575#issuecomment-688644225)
 * Pedestrian mode will no longer respond to "wasd" keys when the user is typing in some input field.
+* Add `options` to CSV papaparsing
 * [The next improvement]
 
 #### 8.0.0-alpha.81